\name{distfreereg-package}
\alias{distfreereg-package}
\docType{package}
\title{
\packageTitle{distfreereg}
}
\description{
\packageDescription{distfreereg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{distfreereg}
\packageIndices{distfreereg}
}
\author{
\packageAuthor{distfreereg}

Maintainer: \packageMaintainer{distfreereg}
}
